Split `query-graphs.js` into `tree-rendering.js` and `app.js`
This commit splits the reusable tree-rendering logic from the
webpage-specific parts.

The `tree-rendering` module still makes a few assumptions which
prohibit its easy reusal (e.g., it assumes that the query-graph
is the only SVG element on the page).
